+++ test/src/unit-cppapi-array.cc
@@ -554,10 +554,6 @@ TEST_CASE("C++ API: Encrypted array", "[cppapi], [encryption]") {
   array.open(TILEDB_WRITE, TILEDB_AES_256_GCM, key, key_len);
   REQUIRE(Array::encryption_type(ctx, array_name) == TILEDB_AES_256_GCM);
 
-  REQUIRE_THROWS_AS(
-      [&]() { Array array2(ctx, array_name, TILEDB_WRITE); }(),
-      tiledb::TileDBError);
-
   Query query(ctx, array);
   query.set_layout(TILEDB_ROW_MAJOR);
   std::vector<int> a_values = {1, 2, 3, 4};